Quick script I wrote to help myself wake up
There is a lot I want to do with is, some of it is somewhat described as TODO's
After a lot of not working correctly alarm apps options that promised to play my songs of choice from Spotify, I decided I will do something a bit more polished on my own. So far it only works if you have installed and running Spotify and installed spotifycli (https://pypi.org/project/spotify-cli-linux/) on Linux, but I want to make it as hastle free as possible, and make it to work on RPi. So just double click, prepare music (so far no clue how to handle that) and go to sleep to be then waken up by your favourite piece